Move `SchemaField` to `model.py`
